[TECHSPIKE] Replace the Modal from NativeBase with ReactNativeModal as the official wrapper on the ActionSheetController
Currently the accessibility support for iOS doesn't work. Setting attributes like `initialFocusRef` doesn't work for other elements beyond TextInputField; Even by marking `accessibilityViewIsModal` for the view, or force to reset the focus in the open modal callback doesn't work.
The solution is to replace the 3rd party library with a top view on a Modal component with the official ActionSheetController. Only by doing this the focus area can be restricted to the actionSheet elements. Note that this is only a temporary fix on iOS side and hasn't been tested with Android accessibility screen reader.

Still has a few issues remaining to be further patched:
1. Animation is not smooth as the original ones
2. Bottom safe area space is hardcoded - using safeArea in Box wrapper is not working.